Remove macros from TypeRegistry
